+++ CHANGELOG.md
@@ -9,6 +9,12 @@ his file's format is define in
 and the release workflow reads it to set github's release notes.
 
 
+## Unreleased
+
+### Fixed
+
+- Use the standard "Authorization" heaer for "Bearer <token>"
+
 ## [0.19.2] 2022-3-7
 
 ### Changed